+++ src/model/locatable.h
@@ -3,16 +3,16 @@
 
 class Locatable { 
     private:
-        float x, y; 
+        int x, y; 
     
     public: 
-        Locatable(float x, float y);
+        Locatable(int x, int y);
 
-        float getX() {return x;} 
-        float getY() {return y;}
-        void setX(float x) {this->x = x;} 
-        void setY(float y) {this->y = y;}
-        void setPosition(float x, float y) {this->setX(x); this->setY(y);}
+        int getX() {return x;} 
+        int getY() {return y;}
+        void setX(int x) {this->x = x;} 
+        void setY(int y) {this->y = y;}
+        void setPosition(int x, int y) {this->setX(x); this->setY(y);}
 
         float distanceTo(Locatable l);
 };
